Replaced old Button component with the new one, updated Button component about: fn is optional, type change to submit
